developed handleOptionClick functionality
