+++ pages/api/studentPortal/quizzes/firstProgrammingLanguage.ts
@@ -1,21 +1,4 @@
-export type QuizQuestion = {
-  title: string;
-  body: string;
-  options: QuizOption[];
-  maxSelections?: number;
-};
-
-export type QuizOption = {
-  name: string;
-  result?: string;
-  weight?: number;
-};
-
-export type QuizData = {
-  title: string;
-  body: string;
-  questions: QuizQuestion[];
-};
+import { QuizData } from "../../../../components/resources/quizzes/shared/types";
 
 export const quizData: QuizData = {
   title: "What Coding Language Should I Learn First?",
